I forgot onbe file

+++ core/backend/admin/user/login/logout.php
@@ -1,13 +1,17 @@
 <?php
 	include('./navi.php');
 	$cookie = $_COOKIE["acess"];
-	$verbindung = mysql_connect ("localhost","db_acess", "raspberry") or die ("keine Verbindung möglich. Benutzername oder Passwort sind falsch");
- 	mysql_select_db("login_sec") or die("db geht nicht");
+	include('../../../../config/connect.db.inc.php');
+	//Check if Database connection established
+	if (mysqli_connect_errno()) {
+		printf("Verbindung fehlgeschlagen: %s\n", mysqli_connect_error());
+		exit();
+	}
  	include('./check_auth.php');
 	if (!$authokay) {
 		die('nobody to logout');
 	}
  	$loeschen = "DELETE FROM cookiedata WHERE cookie_hash = '$cookie'";
-	$loesch = mysql_query($loeschen);
+	$loesch = mysql_query($connection, $loeschen);
 	print('logged out');
 ?>
